Refactored components to allow for greater modularisation - TagSearchFilter refactored into TagSearchFilter, TagSelector, TagSelectorMobile and TagSelectorRegular - This will allow for tag selecting much easier in create-post-form
